__python__/requirements.txt
requests>=2.31.0
urllib3>=2.0.7
python-dateutil>=2.8.2
yfinance>=0.2.36
numpy>=1.26.3
pandas>=2.2.0
matplotlib>=3.8.2
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
flask>=3.0.0
flask-cors>=4.0.0
plotly>=5.18.0
tqdm>=4.66.1 